Moved extra_params float cast to mixin, added test for it
