+++ observation_portal/common/mixins.py
@@ -40,3 +40,16 @@ class CustomIsoDateTimeFilterMixin(object):
             if isinstance(f, IsoDateTimeFilter):
                 f.field_class = CustomIsoDateTimeField
         return filters
+
+class ExtraParamsFormatter(object):
+    ''' This should be mixed in with Serializers that have extra_params JSON fields, to ensure the float values are 
+        stored as float values in the db instead of as strings
+    '''
+    def to_internal_value(self, data):
+        data = super().to_internal_value(data)
+        for field, value in data.get('extra_params', {}).items():
+            try:
+                data['extra_params'][field] = float(value)
+            except ValueError:
+                pass
+        return data
