+++ observation_portal/requestgroups/test/test_api.py
@@ -1294,6 +1294,14 @@ class TestConfigurationApi(SetTimeMixin, APITestCase):
         self.assertEqual(response.status_code, 400)
         self.assertIn('must have at least one instrument configuration', str(response.content))
 
+    def test_extra_params_saved_as_float_not_string(self):
+        good_data = self.generic_payload.copy()
+        good_data['requests'][0]['configurations'][0]['extra_params'] = {'test_value': '1.15'}
+        response = self.client.post(reverse('api:request_groups-list'), data=good_data)
+        self.assertEqual(response.status_code, 201)
+        extra_params = response.json()['requests'][0]['configurations'][0]['extra_params']
+        self.assertEqual(extra_params['test_value'], 1.15)
+
     def test_must_have_exposure_time(self):
         bad_data = self.generic_payload.copy()
         del bad_data['requests'][0]['configurations'][0]['instrument_configs'][0]['exposure_time']
